requirement.txt
numpy
pandas
psycopg2-binary
pyarrow
python-dateutil
python-dotenv
requests
six
SQLAlchemy
typing_extensions
tzdata
urllib3
streamlit